+++ src/app/components/directives/red-directive.ts
