+++ src/app/components/header/header.component.ts
@@ -1,17 +1,17 @@
 import { Component, inject } from "@angular/core";
 import { CommonModule, NgIf } from "@angular/common";
 import { MatToolbarModule } from '@angular/material/toolbar';
-import { CustomRemoveDashes } from "../../pipes/remove-dashes.pipe";
-import { CurrentDateTimePipe } from "../../pipes/current-date-time.pipe";
-import { menuItems } from "../../utils/menu-items";
-import { toLowerCaseArray, toUpperCaseArray } from "../../utils/upper-lower-case-menu-item";
+import { CustomRemoveDashes } from "../pipes/remove-dashes.pipe";
+import { CurrentDateTimePipe } from "../pipes/current-date-time.pipe";
+import { menuItems } from "../utils/menu-items";
+import { toLowerCaseArray, toUpperCaseArray } from "../utils/upper-lower-case-menu-item";
 import { RouterLink } from "@angular/router";
-import { UserService } from "../../service/user.service";
+import { UserService } from "../service/user.service";
 import { AuthenticationComponent } from "../authentication/authentication.component";
 import { MatDialog, MatDialogRef } from "@angular/material/dialog";
 import { Subject, takeUntil } from "rxjs";
 import { logoutComponent } from "../logout/logout.component";
-import { ColorBasketDirective } from "../../directives/color-basket.directive";
+import { ColorBasketDirective } from "../directives/color-basket.directive";
 
 @Component({
     selector: 'app-header',
