+++ src/app/components/pipes/limit-string.pipe.ts
