+++ src/app/components/pipes/remove-dashes.pipe.ts
