+++ src/app/components/pipes/upper-case.pipe.ts
