+++ src/app/components/todo-list/edit-todo-dialog/edit-todo-dialog.component.ts
@@ -6,7 +6,7 @@ import { MatSnackBar } from '@angular/material/snack-bar';
 import { MatFormFieldModule } from "@angular/material/form-field";
 import { MatInputModule } from "@angular/material/input";
 import { MatIconModule } from "@angular/material/icon";
-import { ITodo } from "../../../interface/todo.interface";
+import { ITodo } from "../../interface/todo.interface";
 import { completedValidator } from "../create-todo-form/create-todo-form.component";
 
 @Component({
