+++ src/app/components/users-list/delete-user-dialog/delete-user-dialog.component.ts
@@ -1,5 +1,5 @@
 import { Component, inject } from "@angular/core";
-import { IUser } from "../../../interface/user.interface";
+import { IUser } from "../../interface/user.interface";
 import { MAT_DIALOG_DATA, MatDialogActions, MatDialogModule } from "@angular/material/dialog";
 import { MatButton } from "@angular/material/button";
