+++ src/app/components/utils/upper-lower-case-menu-item.ts
